requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
azure-cosmos==4.5.1
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
openpyxl
xlrd==2.0.1
azure-storage-blob
azure-identity
azure-cosmos
fastapi[all]
fastapi
uvicorn
python-dotenv
python-multipart
azure-search-documents==11.4.0
azure-ai-textanalytics==5.3.0
openai==1.3.0
PyPDF2==3.0.1
python-docx==1.1.0
httpx==0.27.0